pyproject.toml
[tool.poetry]
name = "fastapi-pet-project"
version = "0.1.0"
description = ""
authors = ["Almaz Durusaliev <adurusaliev@megacom.kg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
redis = "^5.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.10.3"}
bcrypt = "3.2.2"
python-multipart = "^0.0.20"
pydantic-settings = "^2.8.1"
uvicorn = "^0.34.0"


[tool.poetry.group.linters.dependencies]
pre-commit = "^4.0.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
